requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.0
Django==4.1.5
django-countries==7.5
django-fontawesome==1.0
django-rest-knox==4.2.0
django-translations==1.3.0
djangorestframework==3.14.0
Pillow==9.4.0
pycparser==2.21
python-gettext==4.1
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3

django-modeltranslation==0.18.7
pip==22.3.1
wheel==0.38.4
setuptools==66.0.0